Admin Middleware

+++ app/Http/Controllers/MessageController.php
@@ -13,9 +13,15 @@ class MessageController extends Controller
         return view('admin.messages.index', compact('messages'), ['title' => 'Messages']);
     }
 
-    public function delete(string $id)
+    public function delete(string $id, Request $request)
     {
+        if (!$request->query('token') || $request->query('token') !== env('ADMIN_ACCESS_TOKEN')) {
+            return redirect('/')->with('error', 'Unauthorized access!');
+        }
+
         Message::destroy($id);
-        return redirect()->back()->with('success', 'Message removed successfully');
+
+        return redirect('/admin/messages?token=' . $request->query('token'))
+            ->with('success', 'Message removed successfully');
     }
 }
